Refactoring Invoice to cleanup unused code
